+++ hashtree.h
@@ -179,7 +179,7 @@ public:
 
     // for transfertest.cpp
     Storage *       get_storage() { return storage_; }
-    void            set_size(uint64_t size) { size_ = size; fprintf(stderr,"hashtree: set_size %llu\n", size ); }
+    void            set_size(uint64_t size) { size_ = size; fprintf(stderr,"hashtree: set_size %llu\n", (long long unsigned int)size ); }
 
     // Arno: persistent storage for state other than hashes (which are in .mhash)
     int serialize(FILE *fp);
@@ -251,7 +251,7 @@ public:
 
     // for transfertest.cpp
     Storage *       get_storage() { return storage_; }
-    void            set_size(uint64_t size) { size_ = size; fprintf(stderr,"zerohashtree: set_size %llu\n", size ); }
+    void            set_size(uint64_t size) { size_ = size; fprintf(stderr,"zerohashtree: set_size %llu\n", (long long unsigned int)size ); }
 };
