+++ sendrecv.cpp
@@ -98,7 +98,7 @@ bin_t        Channel::DequeueHint (bool *retransmitptr) {
     // Arno, 2012-01-23: Extra protection against channel loss, don't send DATA
     if (last_recv_time_ < NOW-(3*TINT_SEC))
     {
-    	dprintf("%s #%u dequeued bad time %llu\n",tintstr(),id_, last_recv_time_ );
+    	dprintf("%s #%u dequeued bad time %llu\n",tintstr(),(unsigned int)id_, (long long unsigned int)last_recv_time_ );
     	return bin_t::NONE;
     }
 
@@ -140,7 +140,7 @@ bin_t        Channel::DequeueHint (bool *retransmitptr) {
     //for(int i=0; i<hint_in_.size(); i++)
     //    mass += hint_in_[i].bin.base_length();
     char bin_name_buf[32];
-    dprintf("%s #%u dequeued %s [%lli]\n",tintstr(),id_,send.str(bin_name_buf),mass);
+    dprintf("%s #%u dequeued %s [%lli]\n",tintstr(),(unsigned int)id_,send.str(bin_name_buf),(long long int)mass);
     return send;
 }
 
@@ -260,7 +260,7 @@ void    Channel::AddHint (struct evbuffer *evb) {
     int allowed_hints = max(0,hints_limit-(int)rough_global_hint_out_size);
 
     if (DEBUGTRAFFIC)
-    	fprintf(stderr,"hint c%d: %f want %d allow %d chanout %llu globout %llu\n", id(), transfer().GetCurrentSpeed(DDIR_DOWNLOAD), first_plan_pck, allowed_hints, hint_out_size_, rough_global_hint_out_size );
+    	fprintf(stderr,"hint c%d: %f want %d allow %d chanout %llu globout %llu\n", (int)id(), (float)transfer().GetCurrentSpeed(DDIR_DOWNLOAD), (int)first_plan_pck, (int)allowed_hints, (long long unsigned int)hint_out_size_, (long long unsigned int)rough_global_hint_out_size );
 
     // Take into account network limit
     uint64_t plan_pck = (uint64_t)min(allowed_hints,first_plan_pck);
@@ -279,8 +279,8 @@ void    Channel::AddHint (struct evbuffer *evb) {
             evbuffer_add_8(evb, SWIFT_HINT);
             evbuffer_add_32be(evb, bin_toUInt32(hint));
             char bin_name_buf[32];
-            dprintf("%s #%u +hint %s [%lli]\n",tintstr(),id_,hint.str(bin_name_buf),hint_out_size_);
-            dprintf("%s #%u +hint base %s width %d\n",tintstr(),id_,hint.base_left().str(bin_name_buf), hint.base_length() );
+            dprintf("%s #%u +hint %s [%lli]\n",tintstr(),(unsigned int)id_,hint.str(bin_name_buf),(long long int)hint_out_size_);
+            dprintf("%s #%u +hint base %s width %d\n",tintstr(),(unsigned int)id_,hint.base_left().str(bin_name_buf), (int)hint.base_length() );
             hint_out_.push_back(hint);
             hint_out_size_ += hint.base_length();
             //fprintf(stderr,"send c%d: HINTLEN %i\n", id(), hint.base_length());
@@ -486,13 +486,13 @@ void    Channel::Recv (struct evbuffer *evb) {
         rtt_avg_ = NOW - last_send_time_;
         dev_avg_ = rtt_avg_;
         dip_avg_ = rtt_avg_;
-        dprintf("%s #%u sendctrl rtt init %lli\n",tintstr(),id_,rtt_avg_);
+        dprintf("%s #%u sendctrl rtt init %lli\n",tintstr(),(unsigned int)id_,(long long int)rtt_avg_);
     }
 
     bin_t data = evbuffer_get_length(evb) ? bin_t::NONE : bin_t::ALL;
 
 	if (DEBUGTRAFFIC)
-		fprintf(stderr,"recv c%d: size %d ", id(), evbuffer_get_length(evb));
+		fprintf(stderr,"recv c%d: size %d ", (int)id(), (int)evbuffer_get_length(evb));
 
 	while (evbuffer_get_length(evb)) {
         uint8_t type = evbuffer_remove_8(evb);
@@ -610,8 +610,8 @@ bin_t Channel::OnData (struct evbuffer *evb) {  // TODO: HAVE NONE for corrupted
 
     // Arno: Assuming DATA last message in datagram
     if (evbuffer_get_length(evb) > hashtree()->chunk_size()) {
-    	dprintf("%s #%u !data chunk size mismatch %s: exp %lu got " PRISIZET "\n",tintstr(),id_,pos.str(bin_name_buf), hashtree()->chunk_size(), evbuffer_get_length(evb));
-    	fprintf(stderr,"WARNING: chunk size mismatch: exp %lu got " PRISIZET "\n",hashtree()->chunk_size(), evbuffer_get_length(evb));
+    	dprintf("%s #%u !data chunk size mismatch %s: exp %lu got " PRISIZET "\n",tintstr(),id_,pos.str(bin_name_buf), (long unsigned int)hashtree()->chunk_size(), evbuffer_get_length(evb));
+    	fprintf(stderr,"WARNING: chunk size mismatch: exp %lu got " PRISIZET "\n",(long unsigned int)hashtree()->chunk_size(), evbuffer_get_length(evb));
     }
 
     int length = (evbuffer_get_length(evb) < hashtree()->chunk_size()) ? evbuffer_get_length(evb) : hashtree()->chunk_size();
@@ -694,7 +694,7 @@ void    Channel::OnAck (struct evbuffer *evb) {
         ri++;
     char bin_name_buf[32];
     dprintf("%s #%u %cack %s %lli\n",tintstr(),id_,
-            di==data_out_.size()?'?':'-',ackd_pos.str(bin_name_buf),peer_time);
+            di==data_out_.size()?'?':'-',ackd_pos.str(bin_name_buf),(long long int)peer_time);
     if (di!=data_out_.size() && ri==data_out_tmo_.size()) { // not a retransmit
             // round trip time calculations
         tint rtt = NOW-data_out_[di].time;
@@ -713,7 +713,7 @@ void    Channel::OnAck (struct evbuffer *evb) {
         if (owd_min_bins_[owd_min_bin_]>owd)
             owd_min_bins_[owd_min_bin_] = owd;
         dprintf("%s #%u sendctrl rtt %lli dev %lli based on %s\n",
-                tintstr(),id_,rtt_avg_,dev_avg_,data_out_[di].bin.str(bin_name_buf));
+                tintstr(),(unsigned int)id_,(long long int)rtt_avg_,(long long int)dev_avg_,data_out_[di].bin.str(bin_name_buf));
         ack_rcvd_recent_++;
         // early loss detection by packet reordering
         for (int re=0; re<di-MAX_REORDERING; re++) {
@@ -1128,7 +1128,7 @@ void Channel::Reschedule () {
         	if (evsend_ptr_ != NULL) {
         		struct timeval duetv = *tint2tv(duein);
         		evtimer_add(evsend_ptr_,&duetv);
-        		dprintf("%s #%u requeue for %s in %lli\n",tintstr(),id_,tintstr(next_send_time_), duein);
+        		dprintf("%s #%u requeue for %s in %lli\n",tintstr(),(unsigned int)id_,tintstr(next_send_time_),(long long int)duein);
         	}
         	else
         		dprintf("%s #%u cannot requeue for %s, closed\n",tintstr(),id_,tintstr(next_send_time_));
